requirements.txt
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.3.0
channels_redis==4.3.0
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==4.2.23
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
pillow==11.3.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
redis==6.2.0
service-identity==24.2.0
simplejwt==2.0.1
sqlparse==0.5.3
tomli==2.2.1
Twisted==25.5.0
txaio==23.6.1
typing==3.7.4.3
typing_extensions==4.14.1
tzdata==2025.2
zope.interface==7.2
